Separated view into different classes for each subelement
